Refactor: Path 수정 및 Alias 적용 #24
