+++ config-overrides.js
@@ -7,6 +7,7 @@ module.exports = override(
   addWebpackAlias({
     '@assets': path.resolve(__dirname, 'src', 'assets'),
     '@components': path.resolve(__dirname, 'src', 'components'),
+    '@containers': path.resolve(__dirname, 'src', 'containers'),
     '@constants': path.resolve(__dirname, 'src', 'constants'),
     '@libs': path.resolve(__dirname, 'src', 'libs'),
     '@recoil': path.resolve(__dirname, 'src', 'recoil'),
