+++ tsconfig.json
@@ -37,6 +37,12 @@
       "@components": [
         "components"
       ],
+      "@containers/*": [
+        "containers/*"
+      ],
+      "@containers": [
+        "containers"
+      ],
       "@constants/*": [
         "constants/*"
       ],
